Added functions in Marker.java to calculate where to place the red and black markers. Updated types in Board.java

+++ Checkers/src/checkers/Board.java
@@ -8,8 +8,8 @@ import java.util.Scanner;
  */
 public class Board {
     
-    double rows;
-    double columns;
+    int rows;
+    int columns;
     Location[][] boardLocations;
     int rowCount;
     int columnCount;
